add route me, route logout, middleware and policy (laravel gates look alike with consise method object func)

+++ app/category/controller.js
@@ -2,6 +2,15 @@ const Category = require('./model');
 
 async function store(req, res, next) {
   try {
+    let policy = policyFor(req.user);
+
+    if (!policy.can('create', 'Category')) {
+      return res.json({
+        error: 1,
+        message: `Anda tidak memiliki akses untuk membuat kategori`,
+      });
+    }
+
     let payload = req.body;
     let category = new Category(payload);
     await category.save();
@@ -21,6 +30,15 @@ async function store(req, res, next) {
 
 async function update(req, res, next) {
   try {
+    let policy = policyFor(req.user);
+
+    if (!policy.can('update', 'Category')) {
+      return res.json({
+        error: 1,
+        message: `Anda tidak memiliki akses untuk mengupdate kategori`,
+      });
+    }
+
     let payload = req.body;
     let category = await Category.findOneAndUpdate(
       { _id: req.params.id },
@@ -43,6 +61,15 @@ async function update(req, res, next) {
 
 async function destroy(req, res, next) {
   try {
+    let policy = policyFor(req.user);
+
+    if (!policy.can('delete', 'Category')) {
+      return res.json({
+        error: 1,
+        message: `Anda tidak memiliki akses untuk menghapus kategori`,
+      });
+    }
+
     let deleted = await Category.findOneAndDelete({ _id: req.params.id });
     return res.json(deleted);
   } catch (err) {
